calculate results with independent filtering
#' This function performs differential expression analysis using DESeq2 after
#' filtering genes to minimize type I error and maximize detection of DEGs.
#'
#' @param dds A DESeqDataSet object.
#' @param lib_index A vector specifying which libraries (columns) are included
#' in the comparison which will be used for filtering based on mapping and the
#' computation of the independent filter statistic theta.
#' @param name_contrast A character vector specifying the contrast or coefficient name.
#' @param contr Logical; if TRUE, use the contrast argument, otherwise use the name argument.
The minimum threshold of normalized mapping events expected
to be present in the libraries included in the comparison. Default is 0, i.e.
Libraries with more than 0 normalized counts
The minimum threshold of libraries with mapping events
above the min_counts value required for a gene to be included. Default is 2,
i.e. Genes included in the comparison have counts above 0 (min_counts) in
more than 2 of the libraries included with lib_index with counts.
Default is "rowVars".